fix/foi arrumado filtro e campo de pesquisa, para facilitar na pesquisa palavras com acento
