+++ frontend/src/Global/GeneralStyled.js
@@ -20,7 +20,7 @@ export const Table = styled.table`
   font-size: 1rem;
   font-weight: 600;
   @media screen and (max-width: 580px) {
-  font-size: 0.9rem;    
+    font-size: 0.9rem;
   }
 `;
 
@@ -32,4 +32,19 @@ export const CellTableTitle = styled.th`
     box-shadow: rgba(50, 50, 93, 0.25) 0px 30px 60px -12px inset,
       rgba(0, 0, 0, 0.3) 0px 18px 36px -18px inset;
   }
+`;
+
+export const Input = styled.input`
+  background-color: #fbfefd;
+  border: #2da77a 1px solid;
+  border-radius: 8px;
+  outline-width: 0;
+  box-shadow: #2da77a 0px 1px 2px, rgba(0, 0, 0, 0.2) 0px 2px 2px 0px,
+    #2da77a 0px -1px 0px inset;
+  font-size: 1.2rem;
+  font-weight: 700;
+  height: 2.6rem;
+  max-width: 52rem;
+  padding: 0 1rem;
+  width: 100%;
 `;
