+++ frontend/src/constants/vowels.js
@@ -1,0 +1,1 @@
+export const vowels = ["aáàâã", "eéèê", "iíìî", "oòóôõ", "uùúû"];
